add nc-tools to cmake
