--- sdk/tools/asset_builder/targets.txt
@@ -1,2 +1,0 @@
-mesh nc/resources/mesh/capsule.fbx
-mesh nc/resources/mesh/sphere.fbx
